Specify number of corrections in add_q
I tried to edit the function to allow the user to specify the number of multiple comparison corrections, as is allowed in the p.adjust function. The optional call would be the same as p.adjust (n = X) and feed directly into the p.adjust function. This change would be especially useful when also using tbl_merge() function as users may want to adjust for the number of models instead of the number of model parameters (ex. adjust for 5 lm models, as opposed to 3 parameters in each lm model). Please feel free to edit the code if I made a mistake, as this is my first pull request ever.